requirements.txt
opencv-contrib-python>=4.8,<5
numpy>=1.26,<3
mss>=9.0,<12
Pillow>=10.0,<13
websocket-client>=1.6,<2
pywin32>=306
pycaw==20251023
sounddevice>=0.4
customtkinter==5.2.2
miniaudio>=1.2
